+++ options/arguments.go
@@ -135,6 +135,31 @@ func (a Argument) String() string {
 	return string(a)
 }
 
+// Is returns true if argument equals to given value
+func (a Argument) Is(value interface{}) bool {
+	switch t := value.(type) {
+	case string:
+		return a.String() == t
+	case int:
+		v, err := a.Int()
+		return v == t && err == nil
+	case int64:
+		v, err := a.Int64()
+		return v == t && err == nil
+	case uint64:
+		v, err := a.Uint()
+		return v == t && err == nil
+	case float64:
+		v, err := a.Float()
+		return v == t && err == nil
+	case bool:
+		v, err := a.Bool()
+		return v == t && err == nil
+	}
+
+	return false
+}
+
 // Int converts argument to int
 func (a Argument) Int() (int, error) {
 	return strconv.Atoi(string(a))